+++ src/main/java/io/leangen/graphql/util/Urls.java
@@ -11,6 +11,7 @@ public final class Urls {
         public static final String NON_UNIQUE_TYPE_NAME = "https://github.com/leangen/graphql-spqr/wiki/Errors#non-unique-type-name";
         public static final String CONFLICTING_RESOLVER_TYPES = "https://github.com/leangen/graphql-spqr/wiki/Errors#operation-with-multiple-resolver-methods-of-different-types";
         public static final String MISSING_ARGUMENT_NAME = "https://github.com/leangen/graphql-spqr/wiki/Errors#missing-argument-name";
+        public static final String UNRESOLVABLE_SUPER_TYPE = "https://github.com/leangen/graphql-spqr/wiki/Errors#unresolvable-super-type";
         public static final String AMBIGUOUS_MEMBER_TYPE = "https://github.com/leangen/graphql-spqr/wiki/Errors#ambiguous-member-type";
         public static final String AMBIGUOUS_PARAMETER_TYPE = "https://github.com/leangen/graphql-spqr/wiki/Errors#ambiguous-method-parameter-type";
         public static final String RELAY_CONNECTION_SPEC_VIOLATION = "https://github.com/leangen/graphql-spqr/wiki/Errors#relay-connection-spec-violation";